refactor: 移除 per-draw scissor 与 IndexType 配置，强制 16-bit 索引
scissor：删除 ScissorRect / enable_scissor / scissor 字段与 draw_scissor()。
scissor 由 BeginRendering 一次设为全幅，是 pass 内不变量，不再逐 draw 下发。
代价：ImGui 不再按窗口裁剪，超出窗口的 UI 内容不会被裁掉。

索引：删除 IndexType 枚举与 resolve_index_type()，绑定点硬编码
VK_INDEX_TYPE_UINT16。HardwareIndexType concept 收窄为只接受 uint16_t，
让漏改的调用点变成编译错误而不是静默索引回绕。

超限 mesh 改在加载期处理，按资产类型分三种策略：

- bgfx .bin（11 个 example）：不拆，只是不再合。VB chunk 的 num_vertices 是
  uint16，每个 group 天然 <=65535，文件里的索引本就是 group 相对的 16-bit；
  旧 loader 把 group_base_vertex 加进索引才溢出。现在索引保持原样，每 group
  一次 draw 用 vertex_offset 补回，与旧路径逐位相同。
  附带修复 build_volume_source：它在 CPU 上消费 mesh.indices 建边邻接，索引
  变 group 相对后必须补回 base_vertex，否则 bunny 第二个 group 静默取错顶点。

- sponza HZMS：重基而非拆分。全局索引跨 190448 顶点，但每 submesh 触及区间很
  窄。rebase_indices_to_16bit() 按三角形推进跟踪 min/max，跨度将达 0x10000 时
  切段，索引写 idx-chunk_min、vertex_offset=chunk_min。实测 25 submesh 得 25
  个 range，零额外 draw，最大重基索引 24173。

- shadowvolumes identity IB：262144xu32 塌缩为 65535xu16，靠步进 vertex_offset
  分段复用。取 65535 而非 65536 是因为 65535=3x21845 可被 3 整除，段边界必然
  落在三角形边界上。

验证：sponza 重基用离线脚本对全部 786723 个索引断言
vertex_offset+out16[i]==原始 indices[i] 且全覆盖无重叠；bgfx 组为构造性等价；
24 个 example 各跑 160 帧全部 exit 0，frame hash 稳定。注意验证层看不见越界
索引（静默 GPU 行为非 API 违规），故以离线证明为主要保障。

+++ examples/example_bump/example_bump.cpp
@@ -165,7 +165,7 @@ std::vector<BumpVertex> build_cube_vertices()
     return vertices;
 }
 
-const std::vector<uint32_t> cube_indices = {
+const std::vector<uint16_t> cube_indices = {
     0, 2, 1, 1, 2, 3,
     4, 5, 6, 5, 7, 6,
     8, 10, 9, 9, 10, 11,
@@ -174,7 +174,7 @@ const std::vector<uint32_t> cube_indices = {
     20, 21, 22, 21, 23, 22,
 };
 
-void calc_tangents(std::vector<BumpVertex>& vertices, const std::vector<uint32_t>& indices)
+void calc_tangents(std::vector<BumpVertex>& vertices, const std::vector<uint16_t>& indices)
 {
     std::vector<glm::vec3> tan_accum(vertices.size(), glm::vec3(0.0f));
     std::vector<glm::vec3> bitan_accum(vertices.size(), glm::vec3(0.0f));
